use openai for chat, direct for embedding
