+++ convex/_generated/api.d.ts
@@ -52,7 +52,6 @@ import type * as util_geometry from "../util/geometry.js";
 import type * as util_isSimpleObject from "../util/isSimpleObject.js";
 import type * as util_minheap from "../util/minheap.js";
 import type * as util_object from "../util/object.js";
-import type * as util_ollama from "../util/ollama.js";
 import type * as util_openai from "../util/openai.js";
 import type * as util_sleep from "../util/sleep.js";
 import type * as util_types from "../util/types.js";
@@ -106,7 +105,6 @@ declare const fullApi: ApiFromModules<{
   "util/isSimpleObject": typeof util_isSimpleObject;
   "util/minheap": typeof util_minheap;
   "util/object": typeof util_object;
-  "util/ollama": typeof util_ollama;
   "util/openai": typeof util_openai;
   "util/sleep": typeof util_sleep;
   "util/types": typeof util_types;
