+++ convex/agent/embeddingsCache.ts
@@ -1,10 +1,8 @@
 import { v } from 'convex/values';
 import { ActionCtx, internalMutation, internalQuery } from '../_generated/server';
 import { internal } from '../_generated/api';
-import * as openai from '../util/openai';
-import { ollamaFetchEmbedding } from '../util/ollama';
 import { Id } from '../_generated/dataModel';
-import { LLM_CONFIG } from '../constants';
+import { fetchEmbeddingBatch } from '../util/openai';
 
 const selfInternal = internal.agent.embeddingsCache;
 
@@ -28,13 +26,7 @@ export async function fetchBatch(ctx: ActionCtx, texts: string[]) {
   if (cacheResults.length < texts.length) {
     const missingIndexes = [...results.keys()].filter((i) => !results[i]);
     const missingTexts = missingIndexes.map((i) => texts[i]);
-    const response = LLM_CONFIG.ollama
-      ? {
-          embeddings: await Promise.all(
-            missingTexts.map(async (t) => (await ollamaFetchEmbedding(t)).embedding),
-          ),
-        }
-      : await openai.fetchEmbeddingBatch(missingTexts);
+    const response = await fetchEmbeddingBatch(missingTexts);
     if (response.embeddings.length !== missingIndexes.length) {
       throw new Error(
         `Expected ${missingIndexes.length} embeddings, got ${response.embeddings.length}`,
