+++ package.json
@@ -25,7 +25,6 @@
     "dotenv": "^16.1.4",
     "eslint": "8.42.0",
     "hnswlib-node": "^1.4.2",
-    "langchain": "^0.0.175",
     "pixi-viewport": "^5.0.1",
     "pixi.js": "^7.2.4",
     "react": "18.2.0",
